requirements.txt
bcrypt==3.2.0
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
idna==2.10
intervaltree==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
paramiko==2.7.2
psutil==5.8.0
pwn==1.0
pwntools==4.4.0
pycparser==2.20
pyelftools==0.27
Pygments==2.8.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
ROPGadget==6.5
six==1.15.0
sortedcontainers==2.3.0
termcolor==1.1.0
unicorn==1.0.2rc3
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
